backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
anthropic>=0.40
firecrawl-py>=1.0
aiosqlite>=0.20
python-dotenv>=1.0
httpx>=0.27
